FastAPI_MongoDB/requirements.txt
fastapi[standard]
pydantic
motor
python-dotenv